requirements.txt
algoliasearch==1.20.0
algoliasearch-django==1.7.0
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.2
Django==2.1.7
django-rest-swagger==2.2.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
MarkupSafe==1.1.1
openapi-codec==1.3.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
PyYAML==5.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
